+++ public/index.html
@@ -15,7 +15,7 @@
     <div id="container"> 
 
     </div>
-    
+    <script src="../public/js/logic.js" type="application/javascript" ></script>
     <script src="../public/js/request.js" type="text/javascript"></script>
 </body>
 </html>
